chore(types): align score and presenter type annotations

+++ src/boardcomposer/solver/maxrects/scoring.py
@@ -1,7 +1,7 @@
 from boardcomposer.solver.maxrects.state import MaxRectsState
 
 
-def score_state(state: MaxRectsState) -> tuple[int, int, int]:
+def score_state(state: MaxRectsState) -> tuple[int, float, float]:
     return (
         len(state.placements),
         -state.packer.width_mm,
